requirements.txt
Django==3.2.9
djangorestframework==3.12.4
markdown==3.3.4
django-filter==21.1
django-cleanup==5.2.0
psycopg2-binary==2.9.1
gunicorn==20.1.0